# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/v1/demo": {
            "get": {
                "description": "get demo data with page",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "demo"
                ],
                "summary": "list",
                "parameters": [
                    {
                        "description": "post params",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/demo.demoListReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "request content",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/request.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "allOf": [
                                                {
                                                    "$ref": "#/definitions/demo.demoListRes"
                                                },
                                                {
                                                    "type": "object",
                                                    "properties": {
                                                        "list": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/definitions/demo.demoListItem"
                                                            }
                                                        }
                                                    }
                                                }
                                            ]
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "description": "create demo data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "demo"
                ],
                "summary": "create",
                "parameters": [
                    {
                        "description": "post params",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/demo.demoCreateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "request content",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/request.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/demo.demoCreateRes"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/api/v1/demo/{id}": {
            "put": {
                "description": "update demo data by id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "demo"
                ],
                "summary": "update",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "post params",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/demo.demoUpdateReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "request content",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/request.Resp"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/demo.demoUpdateRes"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "delete": {
                "description": "delete demo data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "demo"
                ],
                "summary": "delete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "request content",
                        "schema": {
                            "$ref": "#/definitions/request.Resp"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "demo.demoCreateReq": {
            "type": "object",
            "required": [
                "email",
                "name"
            ],
            "properties": {
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "name": {
                    "description": "姓名",
                    "type": "string"
                }
            }
        },
        "demo.demoCreateRes": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "insert id",
                    "type": "string"
                }
            }
        },
        "demo.demoListItem": {
            "type": "object",
            "properties": {
                "avatar": {
                    "description": "头像",
                    "type": "string"
                },
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "id": {
                    "description": "id",
                    "type": "string"
                },
                "name": {
                    "description": "姓名",
                    "type": "string"
                }
            }
        },
        "demo.demoListReq": {
            "type": "object",
            "required": [
                "page",
                "pageSize"
            ],
            "properties": {
                "page": {
                    "description": "页码",
                    "type": "integer"
                },
                "pageSize": {
                    "description": "页大小",
                    "type": "integer"
                }
            }
        },
        "demo.demoListRes": {
            "type": "object",
            "properties": {
                "list": {
                    "description": "数据列表",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/demo.demoListItem"
                    }
                },
                "total": {
                    "description": "总数",
                    "type": "integer"
                }
            }
        },
        "demo.demoUpdateReq": {
            "type": "object",
            "required": [
                "email",
                "id",
                "name"
            ],
            "properties": {
                "email": {
                    "description": "邮箱",
                    "type": "string"
                },
                "id": {
                    "description": "id",
                    "type": "string"
                },
                "name": {
                    "description": "姓名",
                    "type": "string"
                }
            }
        },
        "demo.demoUpdateRes": {
            "type": "object",
            "properties": {
                "id": {
                    "description": "insert id",
                    "type": "string"
                }
            }
        },
        "request.Resp": {
            "type": "object",
            "required": [
                "err_code",
                "err_msg"
            ],
            "properties": {
                "data": {
                    "description": "正常返回数据存放位置"
                },
                "err_code": {
                    "description": "状态码 0 success 1 failed",
                    "type": "integer"
                },
                "err_msg": {
                    "description": "错误信息描述",
                    "type": "string"
                }
            }
        }
    }
}